move semantics in match
matching against a value will do same for all values of a variant
which could result in partial move if the value of the variant doesn't
implement Copy trait.

matching against reference will do same for values of variant so in that
case, the variable matched against will still be available afterwards.