intent communication with lifetimes
over here, we want the output of longest to be valid as long as
x and y are valid. After specifying this intent, the compiler ensures
that it is enforced. If you want a different behaviour, then you have
to change the intent, otherwise you have to code in a way that your
intent is satisfied.